renderer - move rendering utils to appropriate packages
