+++ packages/ui/src/components/Footers/index.ts
@@ -1,4 +1,4 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-export * from './LinksFooter';
+export * from './BaseLinksFooter';
